Update core framework for EDRM XML files
Adds writing the Fields and Documents (Entries) for the EDRM XML to file.  Includes tests to ensure the fields and document nodes are written properly, however these are not really unit tests now as they are not self-contained (they require a human to review the file for correctness.

Also created the EDRMUtilities.py module to isolate some of the common functionality for more general re-use.  Pushed a lot of the repeated XML generation code up to the EntryInterface class, and made DirectoryEntry a subclass of FileEntry to reduce repeated code.