Introduce utility configurations and enhance EDRM architecture
Added a utility configurations dictionary in EDRMUtilities.py which centralizes formatting and default values for various operations. Made refinements in the core EDRM framework, such as isolating common XML generation code to EntryInterface class and simplifying directory entries as a subclass of file entries. These changes enhance the maintainability and efficiency of the system.