Create a sample CSV implementation as an extension to the FileEntry.  Centralize configurations for edrm and data type packages.  Serialize output of EntryInterface.
The new version implements a dictionary of utility configurations in EDRMUtilities.py, simplifying the access and use of formatting and default values across the system. Concurrently, the EDRM framework has undergone a significant refactor, with XML generation code centralized in EntryInterface class and a streamlined inheritance between directory and file entries. These changes strive to improve maintainability and efficiency of the overall system.