Add support for PsListSimple class and update file handling
This commit introduces a new class in nli_tests.py, PsListSimple, to manage CSVRowEntries more effectively. Improvements have also been made to file handling for MappingEntry within nli_generator.py - an additional validation check ensures that no attempt is made to write to a file if the text variable is None or empty. There are also changes to directory, file entry classes and settings related to the CSV editor.